Initial work on generic DSL engine
